+++ ModCore/Utils/CustomDiscordMessageConverter.cs
