+++ ModCore/Utils/EntityFramework/EfPropertyDefinition.cs
