+++ ModCore/Utils/Extensions/Sanitization.cs
