+++ ModCore/Utils/ModCoreSnowflake.cs
