+++ ModCore/Utils/Overwrite.cs
